requirements.txt
langchain
streamlit
plotly
openai
jupyter
python-dotenv
langchain_openai
pydantic
langchain-anthropic
langchain-groq